reorganise cdn worker
