+++ image-cdn/src/error.ts
@@ -1,0 +1,6 @@
+export class ImageNotFoundError extends Error {
+  constructor(message: string) {
+    super(message);
+    this.name = "ImageNotFoundError";
+  }
+}
