+++ image-cdn/src/types.ts
@@ -1,0 +1,6 @@
+export type ImageType = "google_drive";
+export type ImageSize = "small" | "large" | "full";
+export enum ImageSizes {
+  "small" = 400,
+  "large" = 800,
+}
