+++ image-cdn/src/utils.ts
@@ -1,0 +1,3 @@
+export const assertUnreachable = (x: never): never => {
+  throw new Error(`Didn't expect to get here with ${x}`);
+};
